// legacy_mobile_ionic/src/app/models/i-input-options.ts
export interface IInputOptions
{
    /**
     * Lista con las opciones
     */
    List: any;
    /**
     * Valor que se seteara en el *FormControl*
     */
    DataPropsNames: string[];
    DataPropsSepartor: string;
    /**
     * Texto que aparecera en la opcion del Action Sheet
     */
    DescriptionPropName: string;
    /**
     * Titulo del Action Sheet
     */
    ActionSheetHeader: string;

}